% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrafontdb.r
\docType{package}
\name{extrafontdb-package}
\alias{extrafontdb-package}
\alias{extrafontdb}
\alias{package-extrafontdb}
\title{extrafontdb: Package for Holding the Database for the Extrafont Package}
\description{
It is meant to be used with the extrafont package. The extrafont package contains the code to install and use fonts, while the extrafontdb package contains the font database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fbertran/extrafontdb}
  \item Report bugs at \url{https://github.com/fbertran/extrafontdb/issues}
}

}
\author{
\strong{Maintainer}: Frederic Bertrand \email{frederic.bertrand@lecnam.net} (\href{https://orcid.org/0000-0002-0837-8281}{ORCID})

Authors:
\itemize{
  \item Winston Chang
}

}
\keyword{internal}
