\docType{package}
\name{extrafontdb}
\alias{extrafontdb}
\alias{extrafontdb-package}
\alias{package-extrafontdb}
\title{extrafontdb package}
\description{
  Ths package is meant to be used with the extrafont
  package. The extrafont package contains the code to
  install and use fonts, while this package contains the
  directory structure to hold the font database. See the
  \url{https://github.com/wch/extrafontdb} and
  \url{https://github.com/wch/extrafont} for more
  information.
}

