\docType{package}
\name{fontsdb}
\alias{fontsdb}
\alias{fontsdb-package}
\alias{package-fontsdb}
\title{fonts package}
\description{
  Ths package is meant to be used with the fonts package.
  The fonts package contains the code to install and use
  fonts, while this package contains the directory
  structure to hold the fonts database. See the
  \url{https://github.com/wch/fontsdb} and
  \url{https://github.com/wch/fonts} for more information.
}

